fix(validation): enhance input validation and type casting for order and address parameters

+++ controllers/front/refund.php
@@ -24,7 +24,7 @@ class CiklikRefundModuleFrontController extends ModuleFrontController
             );
         }
 
-        $order = new Order(Tools::getValue('orderId'));
+        $order = new Order((int) Tools::getValue('orderId'));
 
         $transactionData = $this->getTransactionData($order);
